Add Bittide project skeleton
This has been stripped from PR #7, to remove the noise from that PR and
to make sure we don't have to run all the formal tests every time we
apply a review suggestion.

Note: the tests on CI rebuild the libraries unnecessarily, but I
couldn't get artifact passing to work properly.