Lower number of Hedgehog tests in Bittide
A while ago we raised the default test limit from 100 to 10_000 to make
catching errors more likely. However, we discovered that this was mostly
caused by poor random number generation strategies - an issue we've since
fixed. We can therefore lower the number of tests run again, significantly
lowering the time it takes to execute tests.