Add `shell.nix`
With the addition of VexRisc (see #198) we're now relying on quite a
number of different tool(chains): C, Python, Haskell, Rust, Scala,
and Verilator. This many toolchains are hard to keep in sync across
different machines hence the introduction of Nix.

To drop to a shell with all the toolchains available, execute:

    nix-shell --pure

Drop `--pure` if you need access to system executables and libraries.