Upgrade Clash for updated Xilinx clock wizard
Clash recently fixed the Xilinx clock wizard, and added support for
instantiation of Xilinx clock generation primitives without manually
using the Vivado Clocking Wizard. This can be used for bittide
instances which are meant to be put on hardware, such as
`clockControlDemo0` and `clockControlDemo1`.

Only the packages `clash-prelude` and `clash-lib` needed to be
updated for these changes, but for consistancy, all packages from the
`clash-compiler` repository have been updated.