Add bitstream generation to Shake
Vivado needs each output port to be mapped to a pin to generate a
bitstream. This is added in the synthesis step of Shake. Shake will
see if an XDC file with the same name as the target is present in the
`data/constraints` directory. If so, the step to read that XDC file
is added to the generated `run_synth.tcl`.

The default part for synthesis, and thus the steps after that,
remains `xcku035-ffva1156-2-e`, as this part doees not require a
license. The constraint files for the two clock control demos are
for part `xcku040-ffva1156-2-e`, which is the part used in Bittide.

Currently, the part for which a target is synthesized is not checked
before generating the bitstream. This means that bitstream generation
will error if the previous steps were done for a different part.

Note that each step for a target depends on the previous step:
`hdl -> synth -> place -> route -> netlist -> bitstream`

Issue: #267