Add support board specific IP generation
Add support for generation of device specific IP. We already supported for
setting a `part` using the environment variable `SYNTHESIS_PART`. We now
also support setting a board part with the `SYNTHESIS_BOARD`. An error is
thrown when both are set. If neither is set the default part is still
used (`xcku035-ffva1156-2-e`).