Add systemd service file for Vivado Hardware Server
Since the hardware-in-the-loop tests on CI depend on the Vivado Hardware
Server, the Gitlab runner should not start before it. We therefore add a
systemd service which starts the Vivado Hardware Server. That service must run
before starting the gitlab runner service.

Since the "unit" file for the gitlab runner is generated by another script, we
add an additional configuration for the gitlab runner service with the
`systemd drop-in` mechanism.