Fix synthesis for instances without IP
PR #325 added support for setting a board part. However, this change
let to an issue where bittide instances without IP would not
synthesize. This issue did not show up in the CI jobs for staging,
but it did for our nightly runs. The issue is fixed by only creating
and reading IP if the instance has IP.