Fail `FullMeshHwCc` if transceivers go down during CC
We've seen link go down if clock control changes clock speeds too
quickly. However, this would go unnoticed as the transceiver code is
written such that it re-establishes a link after it goes down. This is
usually what we want for resilient design, but in this case it just
obfuscates real issues. This patch therefore makes the test fail if a
link goes down during clock control.