Extend `FullMeshHwCc` with RiscV copy strategy
PR #308 added a full mesh clock control demo instance (+test), where the
clock control (Callisto) was implemented directly in FPGA fabric. In the
end, we want to move this to software based clock control.

This commit introduces a stepping stone towards that: it exposes
decisions from hardware clock control to a register readable by a RiscV
core. The RiscV core then copies this to FINC/FDEC.

Co-authored-by: Martijn Bastiaan <martijn@qbaylogic.com>
Co-authored-by: Lara Herzog <lara@qbaylogic.com>