Write out vivado exit code for caching purposes
Shake now writes the exit code of a hardware in the loop test to a file, so
Shake knows it already performed the test. For now this is only useful when
used locally, but comes in handy when post processing of ILA data is added.
To force a rerun of a HITL test use the flag `--force-test-rerun`.