Add selective synth and test CI runs for debugging
While debugging, we often only want one bittide instance to be tested with our
hardware-in-the-loop infrastructure. With the increasing number of bittide
instances with are synthesized and tested, these CI runs take a long time. With
this commit you can add a file `.github/synthesis/debug.json`, with only the
instances you want CI to synthesize/test. The CI run will always fail on the
'all' job when this file exists to prevent a premature merge.