Don't reset start_probe after HITL test finished (#423)
We currently reset a `start_probe`s value after a HITL test. This means that
after all sub-tests in a HITL test are completed (and successful) all
`start_probe` values are 0. We want to use GDB in post processing, and so we
want the last test to remain active. This commit resets all `start_probe`s
before starting each sub-test, so the last `start_probe` remains active after
all tests complete. This is not ideal, and we would like to update HITL tests
to call external binaries before starting a sub-test.