Fix asymmetric buffer occupancies (#453)
#415 changes the ILA dump structure for the elastic buffer sizes to
only report differences since the last and the current capture in
order to compress the amount of data being reported.

It may be that a difference grows too fast and exceeds the compressed
data-size-boundary between two captures, which is why a fail-safe
procedure was implemented creating some additional capture before that
will happen (resetting the difference counter). However, this
protection mechanism was not setup correctly. As a consequence, the
compressor sometimes ends up in a situation, where it starts measuring
a new difference although the previous difference never got
reported. Hence, it's accumulated value got lost, which caused the
plotted graphs to miss some amount of the buffer changes resulting in
the asymmetric plots.

The update fixes the correct handling of the overflow-protection
trigger and the difference output creation of the data compressor.