Run Docker in priviliged mode with host mappings on HITL runner
`--privileged` is needed to access USB devices (i.e., JTAG), while
`--usersns host` is needed to get the former to work. This is done as
part of a separate PR to give everyone the chance to rebase on top of
it, as running CI jobs with and without the flags causes permission
errors as the different docker instances will write files under
different users.