Add FInc/FDec test for FMC clock chip
This test should mirror FincFdecTests.hs but in software. Since HITL tests use
a VIO, the current test is communicated to the CPU using a control register. The
CPU communicates the status of the test back to the hardware using a status
register, so it can be captured by the VIO.

The frequency adjustments are slower over I2C compare to the dedicated pins on
the external clock board (Si5395J-A-EVB). To speed up the test, the step size
changed from 1ppb to 1ppm.