Correct hourglass argument order (#532)
Fixes the hourglass topology, which is just a special case of a dumbbell with
no rod and balanced weights. However, these configuration parameters were
passed in the wrong order resulting in a pendulum topology with the weight
equals to the thread length instead. This PR puts the arguments into the order
right again.