Add `cpu-include-binaries` and `sim-baud-rate` flag
`cpu-include-binaries` includes the corresponding riscv binary(if there is one) in the compilation process. This way we don't need to program the cpu over JTAG.

`sim-baud-rate` sets the baud rate of the uart component to the maximum value allowed by its clock domain. This makes for faster simulation of softcore CPUs