Use `log` instead of manual debugging option
The 'features' of the log module set in the respective Cargo.toml are global.
Adding other features in two Cargo.toml files in the same module gives a
duplicate features error.

Using the log module is slower than the previous method of a global constant
boolean `DEBUG_MODE` with if-statements around `uwrite` calls. Therefore, I had
to increase the values for when to start reading/writing in the switch demo PE
test.