Use two versions of openocd.
openocd-vexriscv is required because openocd-riscv does not allow the CPU to be reset while openocd is connected. This is unfortunately required for the current state of swCcTopologyTest and is non-trivial to fix.

We intend to at some point remove this test case and get rid of the custom openocd version in favor of the "official" riscv version.