Remove Rust dependency from all instances
Replaces uses of `memBlobsFromElf` with `vecsFromElf`, which is a start for
making all instances independent of Rust.

Note that several instances had `iSize` or `dSize` which did not match the sizes
specified in their respective memory.x files. The new `IMemWords` and
`DMemWords` are now in line with their memory.x counterpart.