Remove dependency on bitstream from Shake targets program and test
Although programming and hardware testing still need a bitstream (and probes
file), their Shake targets no longer depend on it. With this change, we don't
need to cache the shake.database and shake.lock files, which was a nasty hack.

This does mean users now have to manually generate a bitstream before testing.