Use `freeze` in demo switch's clock control
This makes sure that all buffer occupancies are captured at the exact
same time, instead of "when the CPU gets to it". `Freeze` doesn't know
about active links, hence the software needs to more defensive in which
buffers it reads out and which it doesn't.

Note that we don't switch over `swcctopologies` -- the idea being that
that instance will get entirely replaced by software based sample
capturing.

This is part of #742.