Remove ILA sample storage in favor of SW RAM storage
In #899 we introduced a way of storing clock control results in (block)
RAM. In that PR, we add it to the switch demo, which is a DUT that
closely resembles what we think a "real" bittide system might look like.
Due to difficulties with tying the "new" way together with the "old"
way, it leaves the pre-existing method for what it is. This has left us
with a split in how we store clock control data. This commit completely
removes the "old" way of doing things -- in a regrettably large diff.

To summarize:

* It unifies `SwCcTopologies` and `SwitchDemo`, the former now being an
  alias of the latter. In the future we should update our test
  infrastructure such that a single DUT can have multiple test suites
  associated with it.

* It moves all parsing to `Bittide.Instances.Hitl.Driver.Clock.Control`
  and plotting logic a directory below (`Plot`). A single script
  `bittide-instances:samples-to` has been added to deal with the data
  dumps generated by GDB. It can create CSVs, individual clock and
  elastic buffer occupancy plots, and reports.

* It simplifies CI by storing what jobs generate data suitable for clock
  control reports in the JSON.

Regrettably, it also removes a bunch of features that the previous
topology tests used to have: topologies, clock offsets, and startup
delays. The plan is to reintroduce these in later PRs.

Lastly, the plots make it obvious that the current design of the switch
demo is lacking in some ways:

 * Domain difference counters start as soon as all links are stable,
   which means clock control will never see them starting at zero. This
   isn't a problem in practice as clock control is a simple
   P-controller, but it makes the plots ugly / counter-intuitive.

 * Clock control and the associated debug infrastructure runs in the
   _free_ domain. This means that the clock plots show measurements of
   _uncontrolled_ clocks, which is obviously not what we want.