Remove cyclic dependency from the switch demo test.
This is accomplished by moving the components from `Hitl/SwitchDemo.hs` that
produce a memory map into a new file, `Hitl/Dut/SwitchDemo.hs`, and leaving
everything else in `Hitl/SwitchDemo.hs` that doesn't produce a memmap in the
original file. That then changes the dependency graph from

Driver ----+
  ^        |
  |        v
  +------ Hitl

Which is cyclic, into

Driver ----+
 ^         |
 |         v
Dut ----> Hitl

Which is acyclic.