Don't start sending user data without `rxReady`
The transceiver logic implements backpressure in both directions through
two inputs: `rxReady` and `txReady`. The former controls signalling to
the other side whether it is ready to receiver user data, while the
latter signals to the logic that it is ready to supply user data. Before
this patch, the logic would only wait for `txReady` and the `rxReady`
from the *other* side before sending data. This doesn't work, because
transitioning to sending user data means losing the ability to supply
backpressure using *its own* `rxReady`. After this patch, whether or not
the transceiver logic starts sending data also depends on its own
`rxReady`. I.e., if the user of the transceiver logic never indicates
it is ready to receive user data, it will also never starts sending it.