Use packed data in register
This makes sure the bus can write intermediate values to the register,
without the register blowing up because the intermediate value does not
have a representation in the target type. For example, say we have a
color type:

    data RGB = R | G | B

The data type can represent three values, but its bit width is 2 bits.
This necessarily means that one of the bit values is invalid, in this
case 0b11. Now say we store this in an `Either RGB (BitVector 2)`. If
the current value of the register is `Right 0b11`, just writing the
constructor tag (a perfectly valid thing to do from the CPUs point of
view) will result in `Left (unpack 0b11)`, an invalid value!

Note that this strategy might still expose the circuit to intermediate
(invalid) values. I.e., these situations still call for custom locking
logic.