Implement atomic registers
This commit extends the current `registerWb` with an (optional) lock
register that can be used by the software to either atomically read or
write from/to a register. By default, the register decides to include a
lock register or not by reasoning about whether the software would (in
practice) write to it in multiple bus accesses or not. Users can opt-out
of a lock register by passing `lock = NoLock` to `registerConfig`.