Add DNA to `gppe`
Only one `DnaPortE2` can be instantiated in a design, so we now instantiate it
at the top-level and distributed to all CPUs which need it.

Technically the `switchDemo` with a hardware processing element instead of a
`gppe` only has the managementUnit CPU which needs the DNA, but I've also
applied the change there to match the other 2 demo's.