Add abstraction trait over calendar types
Previously abstractions were only made over a single switch calendar, however
those abstractions should be useable for all other calendar types. As such, it
was necessary to make those abstractions agnostic over what calendar they were
used on. Now to use the HAL for calendars, one only needs to provide an
implementation of the `CalendarInterface` trait. In order to facilitate this,
a helper macro is provided (`impl_calendar_interface`). For this I also make
use of some wrapper traits around `as` casts, which are both sealed with a
supertrait.