Use separate JTAG IDs in multi-CPU designs
This commit adds 3 CPUs (for 4 in total) to `bittide-cpus`. Each of
these CPUs is configured with a slighly different JTAG ID:

      /-- Count
      | /-- Part number (ASCII 'QB', the first two letters of QBayLogic)
      | |    /-- Manufacturer (ID for "unregistered")
      v v    v
    0x0_514C_001

We increase the _count_ to differentiate between different CPUs in the
system. In practice, this amounts to finding the relevant `jtagChain`
instantiation in your design and enumerating left-to-right. For example,
when you have:

    [muJtag, ccJtag, gppeJtag] <- jtagChain -< jtag

The management unit would get ID `0x0_514C_001`, the clock control CPU
would get `0x1_514C_001`, and the general purpose processing element
would get `0x2_514C_001`. This isn't automatic though: the `PeConfig`s
associated with each CPU still need to instantiate the right CPU from
`Bittide.Cpus.Riscv32imc`.

This change is motivated by the fact that OpenOCD, in contrast to its
documentation, does not reliably assign GDB ports in chain order. By
making each CPU have a different ID we can pick them out. Ideally, this
would again be done by OpenOCD, making it produce stable GDB port
numbers in lieu of doing it automatically based on chain order. Sadly,
we don't live in this world, so we do it ourselves by inserting a bit of
TCL at the end the chain configuration:

    set tap_list [jtag names]
    foreach tap $tap_list {
        set idcode [jtag cget $tap -idcode]
        echo "JTAG_ID: $tap: [format 0x%08X $idcode]"
    }

We then parse this and connect to the right GDB based on IDCODE (i.e.,
chain order).