Generalize MU for SwitchDemo
The SwitchDemo was the last demo which used a specific MU. It is now generalized
by once again copy-pasting `managementUnit`, but all 3 versions are identical
except for a hardcoded type for the number of internal/external busses. This
should make it easier to generalize the CPUs further and move them to the
`bittide` module (as was intended with `node`).